+++ releaseNotes.md
@@ -7,6 +7,7 @@ edge / 2019-05-01
 - Breaking: removed ability in grammar files to have a wordType and keyword share the same name
 - Breaking: getGraph now requires a uniqueId column. Throws if you attempt to extend a non-unique id
 - Breaking: instead of "grammar grammarName" oneliner now use the grammar name keyword
+- Breaking: removed parseWith cellType property
 - Fix: all word types now have default regex of [^ ]* so no need to specify it
 - Fix: grammar code cleanup
 - Fix: small fixes to grammar sandbox